+++ kernel/src/arch/x86_64/ELFKernel.cpp
@@ -25,32 +25,35 @@ namespace x86_64_WorldOS {
         data_size   += 4095;
         bss_size    += 4095;
 
-        // pre-map kernel as Read-only and No execute in case there are more unused sections
+        // convert to pages
+        text_size   /= 4096;
+        rodata_size /= 4096;
+        data_size   /= 4096;
+        bss_size    /= 4096;
 
-        for (uint64_t i = (uint64_t)kernel_virt; i < kernel_size; i+=4096) {
-            uint64_t phys_addr = i - (uint64_t)kernel_virt + (uint64_t)kernel_phys;
-            x86_64_map_page_noflush((void*)phys_addr, (void*)i, 0x8000001);
+        // pre-map kernel as Read-only and No execute in case there are more unused sections
+        kernel_size += 4095;
+        kernel_size /= 4096;
+
+        for (uint64_t i = 0; i < kernel_size; i++) {
+            x86_64_map_page((void*)((uint64_t)kernel_phys + (i * 4096)), (void*)((uint64_t)kernel_virt + (i * 4096)), 0x8000001);
         }
 
         // actually do the mapping
-        for (uint64_t i = text_start_addr; i < text_size; i+=4096) {
-            uint64_t phys_addr = i - (uint64_t)kernel_virt + (uint64_t)kernel_phys;
-            x86_64_map_page_noflush((void*)phys_addr, (void*)i, 0x1); // Present, Read-only, Execute
+        for (uint64_t i = 0; i < text_size; i++) {
+            x86_64_map_page((void*)(text_start_addr + (i * 4096) - (uint64_t)kernel_virt + (uint64_t)kernel_phys), (void*)(text_start_addr + (i * 4096)), 0x1); // Present, Read-only, Execute
         }
 
-        for (uint64_t i = rodata_start_addr; i < rodata_size; i+=4096) {
-            uint64_t phys_addr = i - (uint64_t)kernel_virt + (uint64_t)kernel_phys;
-            x86_64_map_page_noflush((void*)phys_addr, (void*)i, 0x8000001); // Present, Read-only, Execute Disable
+        for (uint64_t i = 0; i < rodata_size; i++) {
+            x86_64_map_page((void*)(rodata_start_addr + (i * 4096) - (uint64_t)kernel_virt + (uint64_t)kernel_phys), (void*)(rodata_start_addr + (i * 4096)), 0x8000001); // Present, Read-only, Execute Disable
         }
 
-        for (uint64_t i = data_start_addr; i < data_size; i+=4096) {
-            uint64_t phys_addr = i - (uint64_t)kernel_virt + (uint64_t)kernel_phys;
-            x86_64_map_page_noflush((void*)phys_addr, (void*)i, 0x8000003); // Present, Read-Write, Execute Disable
+        for (uint64_t i = 0; i < data_size; i++) {
+            x86_64_map_page((void*)(data_start_addr + (i * 4096) - (uint64_t)kernel_virt + (uint64_t)kernel_phys), (void*)(data_start_addr + (i * 4096)), 0x8000003); // Present, Read-Write, Execute Disable
         }
 
-        for (uint64_t i = bss_start_addr; i < bss_size; i+=4096) {
-            uint64_t phys_addr = i - (uint64_t)kernel_virt + (uint64_t)kernel_phys;
-            x86_64_map_page_noflush((void*)phys_addr, (void*)i, 0x8000003); // Present, Read-Write, Execute Disable
+        for (uint64_t i = 0; i < bss_size; i++) {
+            x86_64_map_page((void*)(bss_start_addr + (i * 4096) - (uint64_t)kernel_virt + (uint64_t)kernel_phys), (void*)(bss_start_addr + (i * 4096)), 0x8000003); // Present, Read-Write, Execute Disable
         }
 
         return true;